Revert CI build to single stage
Since we no longer test the links of the site on the CI, it does not
make sense to keep multiple build stages, as there is no opportunity to
run multiple stages at the same time. Therefore, to avoid the overhead
of spinning up two containers, I'm merging build and optimization back
into a single stage.

With this simplification, there is no longer a need for a staging
branch, which vastly simplifies the build.